refactor: update test directory
