+++ package.json
@@ -12,6 +12,7 @@
     "@types/react-dom": "16.9.6",
     "@types/react-outside-click-handler": "^1.3.0",
     "@types/react-router-dom": "5.1.3",
+    "@types/react-test-renderer": "^16.9.2",
     "@types/styled-components": "5.0.1",
     "@typescript-eslint/parser": "2.27.0",
     "antd": "4.1.1",
@@ -38,6 +39,7 @@
     "react-router": "5.1.2",
     "react-router-dom": "5.1.2",
     "react-scripts": "3.4.1",
+    "react-test-renderer": "^16.13.1",
     "styled-components": "5.0.1",
     "ts-jest": "25.4.0",
     "typescript": "3.8.3"
@@ -66,7 +68,7 @@
     "lint": "eslint src/**/*.{ts,tsx} --fix",
     "lint:css": "stylelint src/**/*.tsx",
     "build": "env-cmd -f ./.env.production react-scripts build",
-    "test": "react-scripts test --env=jsdom --watchAll=false",
+    "test": "react-scripts test --watchAll=false",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
