+++ src/__tests__/utils/date.test.ts
@@ -1,4 +1,4 @@
-import { formatData, parseTime, parseTimeNoSecond, parseDateNoTime, parseDiffDate } from '../utils/date'
+import { formatData, parseTime, parseTimeNoSecond, parseDateNoTime, parseDiffDate } from '../../utils/date'
 
 describe('Date methods tests', () => {
   it('format date data', async () => {
