+++ src/__tests__/utils/number.test.ts
@@ -5,7 +5,7 @@ import {
   parseIndicator,
   parseEpochNumber,
   parseUDTAmount,
-} from '../utils/number'
+} from '../../utils/number'
 
 describe('Number methods tests', () => {
   it('local number string', async () => {
